Create function to setup db client

+++ products-api/products/domain/models/product.go
@@ -1,10 +1,10 @@
 package models
 
 type Product struct {
-	ID          int    `json:"id"`
-	Name        string `json:"name"`
-	Description string `json:"description"`
-	BarCode     string `json:"barCode"`
-	Price       int    `json:"price"`
-	Stock       int    `json:"stock,omitempty"`
+	ID          int     `json:"id"`
+	Name        string  `json:"name"`
+	Description string  `json:"description"`
+	BarCode     string  `json:"barCode"`
+	Price       float64 `json:"price"`
+	Stock       int     `json:"stock,omitempty"`
 }
